Importation of our new structure
Here are the schedule.c and schedule.h files. For now we only designed
the structure we thought about.  This structure is composed by 4
circulars lists of threads, cores, sockets and numas. Each node of the
structure has a link to his father and an array of link to his sons. It
also has an id.  Eventually, we have 4 lookups arrays of pointer to link
to every threads, cores, sockets, numas.  Next step will be to implement
the functions that take a thread, a core or a socket and also the
fucntion that put the threads back.