Fix up comments
When I explained how we do comments before, I failed to mention that we
only use the '/* ... */' style comments for *actual* comments, and we
use the '// ...' style comments to comment out blocks of unwanted code.

Also, it looks like there was some confusion in what I mean by this:
"""
Comments should have a '*' on each line, and there should be a
space before the initial '/*' and the first word. Likewise, there
should be a space after the last '*/'.
"""

What we actually want is something like:

/* Hi this is a really long comment, so I will need to break onto the
 * next line instead of just fitting one one.  Since I am really long, I
 * need to make sure my comments are formatted properly. */

Not:

/* Hi this is a really long comment, so I will need to break onto the */
/* next line instead of just fitting one one.  Since I am really long, */
/* I need to make sure my comments are formatted properly. */

The comment about "there should be a space after the last '*/'" was just
wrong.  I meant there should be a space between the last word and the
'*/' at the end.

I have fixed these up in the code.