Consolidate the init_* functions into create_nodes
Now that we have the lookup arrays and the node lists indexable by the
node_type enum, we can condense node creation into a single function.

This was previously not possible because these lists were hardcoded with
names like socket, core, numa, etc.  Using enums and arrays makes this
possible.