Don't call request_*_specific, request directly
There is no need to call request_*_specific() on the children
when removing a node. Call request_node() instead.
Also, removing a parent recursively goes through the same
process as removing the current node, so start there instead of at the
parent and remove all the way back to the root.