Use refcount instead of bool when removing a node
This will make it easier to do reinsertions later on.  Both removal and
reinsertion requests come from children, so a reinsertion can simply be the
opposite function of a removal.

Where removals do:
if (refcount == 0)
	remove_from_list();
refcount++;

An insertion can do:
refcount--;
if (refcount == 0)
	reinsert_in_list();

This is much better than trying to do something that tracks the number of
children or walks a bunch of lists to figure out if it's appropriate to
reinsert after a child has been returned.