Add code for yielding a core back to the system
This is based on keeping a proper refcount when removing a node and propogating
the resinsertion request up the hierarchy whenever a core is yielded.  When a
node is requested, it ups the refcount of all of it's ancestors, so that when
it gets reinserted it can down the count.  When the count reaches 0, we know
all children (recursively) have been reinserted into the tree.