Update how topology_info calculated
There was a bug in how chips_per_socket was being calculated.
Previously I was using (1 << core_bits) to directly calculate this.
However, this number is always a power of two and will be broken if the
machine actually has between (1 << core_bits) and (1 << (core_bits - 1))
There for (1 << core_bits) and (1 << chip_bits) should both be
interpreted as "max" values (i.e. rounded up to the next power of two).
the true value of chips_per_socket is then obtained on the fly when we
do our remapping of ids to condense them in our lookup table.

This bug cropped up on a 1 sockets, 12 core machine, with 6 cpus per
socket.