Insertion and update of the field score
We add a field int score to our struct node to keep track of which node we
should better allocate when a "request_any_node" is received. The score is the
sum of the node's refcount and of all its ancestors refcount. This score is
updated in the function update_scores called by find_best_node, which is called
itself when we request_any_node. I am thinking about updating the scores
recursively in the function remove and reinsert node.