Add of a test file and start the tests
Add of test_schedule.c &.h which is a file test to see if what we did on the
schedule file is working.
So far, we allocate and yield all the cores, all the chips, all the sockets and
all the numas.
I also added a cores_per_socket and a cores_per_numa field in the topology.