New versions of update_scores and find_best_node
Thre is no need to walk the linked list.  Just use the arrays themselves
and scan through them.  This is much faster than following a pointer.
In fact, we don't even need the linked lists at all anymore (I mentioned
this in an email).  I will remove them completely in an upcoming commit.

Also, update_scores() needs to be called *after* removing a nodes from
the topology. Previously, this was being called during the
find_best_node() stage, which meant it was out of sync with the newest
state of the tree when the next request came in.  With the old strategy,
it would always be one stat behind.