Remove our node_list completely.
There is no need to maintain a linked list of siblings with our new
strategy for allocating nodes.  It is much faster to scan the array of
nodes at a single level than to walk a linked list of pointers. The
primary advantage of the linked list was to get constant time
performance on just extracting the first element when allocating a node
instead of scanning all available nodes.  Now we are doing the scanning
every time, so there is no need for the linked list.

This removal affected the way the node topology was being printed, so I
just restructured and organized this better as part of this commit.