Rename request->alloc and yield->free
With this data structure we are not actually requesting and yielding
cores.  As Barret mentioned in one of his emails, we are actually
writing something similar to a memory allocator, where we want specific
nodes in our structure to be allocated based on some type of
prioritization strategy (i.e. our score field). Because of this, I have
now renamed all of our API functions to use alloc/free instead of
request/yield to avoid any confusion.