Walk the tree to find best_node based on refcounts
With this new algorithm, we don't need to keep a scoring function.  We
just start at the NUMA node and walk down to the proper level, picking
the node with the highest refcount at each level (so long as this
refcount doesn't indicate that there are no more nodes left at that
level). When we get to our leve, we simply pick any node with a refcoutn
of 0 (which we know will exist based on the path we followed).