Update schedule to work with an array of refcount
Discussing about our algorithm to find the best node, we noticed that we had
no way to be sure the highest refcount node we choose at any level has the
resource we want. For exemple, if I want to allocate a chip. I could choose to
pick it in socket 1 because it has the highest refcount. However, I cannot be
sure that a chip is entirely available in this socket.
To solve this problem, our int refcount is now a 4 entry array (one for each
level). This enable us to know at every level how many nodes are allocate.
For example in the case we have allocated chip0, chip2 and core7, then the numa
node will have numa->refcount[NUMA]=1 (not available), numa->refcount[SOCKET]=1
(one socket allocated in this numa), numa->refcount[CHIP]=3 (three chips
allocated in this numa) and numa->refcount[CORE]=5 (five cores allocated in
this numa). If we look at chip3 refcount, we will only see
chip3->refcount[CHIP]=1 (not available) and chip3->refcount[CORE]=1 (one core
is allocated is this chip).
Here are the changes that enable this algorithm.