Return list of cores when allocating a node
We finally choose to return lists of cores using the STAILQ macros. We need
to be able to add new cores to the list at the tail of the queue.
For now we return a list of cores, of type struct core_list. When we alloc a
node, the function node2list will turn the node that has been choosen to be
allocated into a list of its cores. This function is now working on any node.