Change algorithm of decref_node_recursive
The previous algorithm of decref_node_recursive failed to decrease numa and
socket refcount. The new algorithm update all refcount of the given node.
To do so, it updates the refcount at level i-1 by summing the children's
refcount at this level. Then it simply decrease to 0 at level n->type if the
refcount at level n->type-1 is 0. If the node is a core, then we simply decref
its refcount.
So far all tests are passing. Maybe we can improve the way I coded it.