Improvment of find_best_node
We now return the first node of the type in parameter when there are no other
node of the same type at the level we consider.
For exemple, lets say we have socket0 and socket1 with 4 chips per socket.
On socket0, chip 1,2 and 3 are available (so chip 0 is allocated). On socket1,
chip 0,1,2,3 are available. If we ask for any chip (resp any core), we will
find out at the socket level that no chip (resp core) have ever been allocated
in socket1. Which means, we can return chip0 (core0) because we know it will
be the best.