Cleanup build_topology with smaller helper funcs
Now every step of the build_topology() functions is broken down into
smaller steps that can be reused (even across to build_flat_topology()).
Each step is more logical now, and it is much easier to see what the
logical steps are in building the topology.

It has also been verified on my 32 for c99 machine which has holes in
its apic_id space where it goes form 0-15, then jumps to 32-47.