Update alloc_core to handle provisioning
Let's say we want to alloc_core(n0, p0).
If n0 is provisioned by p0, If n0 is already allocated to p1, we remove the core
from the list of core_owned of p1 and give a new core to p1 (alloc_core_any).
Anyway, give core n0 to p0.