Add reprovision stuff TODO
Not sure we want to do it, but I added the TODO so we remember to speak about
it. Also, here is a sum up of how our code works:

-Proc1 request for a core.
-Try to find the best core for a proc (proc1). We try to first allocate the
best core among those provisioned by proc1.
-If the core we have found (coreX) is provisioned by an other proc (proc2),
we try to find an equivalent core (in terms of distance) and allocate this core.
-TODO: If we can't find an equivalent core, we give coreX to proc1 and try to
reprovision coreX for proc2 by an equivalent or better core (done in alloc_core)
-If we can't find an equivalent or better core, we let coreX provisioned by
proc2 and when proc2 will request coreX, we will give it to him and give proc1 a
new best possible core.