Struct proc has now alloc/not alloc prov core
We made the changes to the struct proc in order to fit the one in Akaros.
For that, we added two list to replace the provision core list. The first one
contains the provision cores allocated to the proc and the second one, the
provision cores that are not allocated to the proc.